test: Move `MockCommandExecutor` to shared
